Finished algorithm and animation
Finished the algorithm and animation, or at least the algorithm works on this current set of points. Next we should write a function to generate a random set of n points, and do some testing to make sure the algorithm is air tight. Then we can go back and make the animation look/feel prettier.