Created random point generator + fixed some stuff
Created a method that generates n points and puts them into our input file. Made animation a little smoother and fixed a bug. Next we need to handle a degeneracy that I believe is caused by having three points in a line. The current text case has an example of this. After we figure that out, continue testing on more pointsets.